pyproject.toml
[project]
name = "confidence-tom"
version = "0.2.0"
description = "Theory of Mind 信心預測實驗 - 大模型能否讀懂小模型的內在信心狀態"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.5.0",
    "fastparquet>=2025.12.0",
    "huggingface-hub>=1.5.0,<2.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "numpy>=1.26.0",
    "tenacity>=9.1.4",
    "tqdm>=4.67.1",
    "transformers>=5.5.0",
    "wandb>=0.24.0",
    "sympy>=1.14.0",
    "antlr4-python3-runtime>=4.9.3",
    "scikit-learn>=1.7.2",
    "pexpect>=4.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/confidence_tom"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
strict = true
warn_unused_configs = true
warn_return_any = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
]
viz = [
    "matplotlib>=3.10.8",
    "seaborn>=0.12.0",
]
local-inference = [
    "accelerate>=1.13.0",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
